LayoutTests/imported/w3c/web-platform-tests/FileAPI/url/url-charset.window-expected.txt


FAIL Blob charset should override any auto-detected charset. assert_equals: expected "UTF-8" but got "windows-1252"
FAIL Blob charset should override <meta charset>. assert_equals: expected "UTF-8" but got "windows-1252"
